server-fastapi/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
celery==5.3.4
redis==5.0.1
qdrant-client==1.6.9
PyPDF2==3.0.1
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
google-generativeai==0.3.2
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
minio==7.2.0